tests: Change a local variable from size_t to int
This fixes the following build warning with MSVC:

    ../tests/generic.c(56): warning C4267: 'function': conversion from 'size_t' to 'const int', possible loss of data

The checkasm_check_* functions currently take the width/height
parameters as int, not size_t.

+++ tests/generic.c
@@ -46,8 +46,8 @@ void checkasm_test_copy(copy_func fun, const char *name)
 
     declare_func(void, uint8_t *dest, const uint8_t *src, size_t n);
 
-    for (size_t w = 1; w <= WIDTH; w *= 2) {
-        if (check_func(fun, "%s_%zu", name, w)) {
+    for (int w = 1; w <= WIDTH; w *= 2) {
+        if (check_func(fun, "%s_%d", name, w)) {
             CLEAR_PIXEL_RECT(c_dst);
             CLEAR_PIXEL_RECT(a_dst);
